+++ data/src/main/java/com/ssing/data/instructorlessondetail/remote/dto/request/InstructorLessonDetailRequest.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class InstructorLessonDetailRequest(
     @SerialName("lessonId")
-    val lessonId: Int,
+    val lessonId: Long,
 )
